add namedMessage and fieldGroup support

+++ entproto/internal/zero/ent/pet_query.go
@@ -374,8 +374,8 @@ func (pq *PetQuery) sqlAll(ctx context.Context, hooks ...queryHook) ([]*Pet, err
 }
 
 func (pq *PetQuery) loadOwner(ctx context.Context, query *UserQuery, nodes []*Pet, init func(*Pet), assign func(*Pet, *User)) error {
-	ids := make([]int, 0, len(nodes))
-	nodeids := make(map[int][]*Pet)
+	ids := make([]uint64, 0, len(nodes))
+	nodeids := make(map[uint64][]*Pet)
 	for i := range nodes {
 		if nodes[i].user_pet == nil {
 			continue
